Fragment ui is created,instant weather module is working

+++ app/src/main/java/com/kaydeniz/weather/Model/Main.java
@@ -3,7 +3,7 @@ package com.kaydeniz.weather.Model;
 public class Main {
 
     private double temp;
-    private int pressure ;
+    private double pressure ;
     private int humidity ;
     private double temp_min;
     private double temp_max;
@@ -19,11 +19,11 @@ public class Main {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public double getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
